starter/requirements.txt
a2a-sdk>=1.0.0
google-adk>=1.31.1
google-cloud-discoveryengine>=0.13.12
toolbox-core>=1.0.0
python-dotenv>=1.2.2